+++ sparkle/playground/src/stories/Pods.tsx
@@ -1,15 +1,15 @@
 import {
   ArchiveIcon,
+  AttachmentIcon,
   Avatar,
   BellIcon,
   BoltOffIcon,
-  BookOpenIcon,
+  Breadcrumbs,
   Button,
   Card,
   ChatBubbleBottomCenterTextIcon,
   ChatBubbleLeftRightIcon,
   CheckDoubleIcon,
-  CodeSlashIcon,
   Cog6ToothIcon,
   ContactsUserIcon,
   Dialog,
@@ -26,10 +26,10 @@ import {
   DropdownMenuSubContent,
   DropdownMenuSubTrigger,
   DropdownMenuTrigger,
+  EyeIcon,
   HeartIcon,
   LightbulbIcon,
   LinkIcon,
-  ListSelectIcon,
   LogoutIcon,
   MoreIcon,
   NavigationList,
@@ -44,30 +44,28 @@ import {
   ScrollArea,
   ScrollBar,
   SearchInput,
-  SidebarLayout,
-  type SidebarLayoutRef,
-  SidebarLeftCloseIcon,
-  SidebarLeftOpenIcon,
   SlackLogo,
   SpaceClosedIcon,
   SpaceOpenIcon,
-  Tabs,
-  TabsContent,
-  TabsList,
-  TabsTrigger,
   TrashIcon,
   UserGroupIcon,
   UserIcon,
   XMarkIcon,
 } from "@dust-tt/sparkle";
-import { useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 
 import { AgentBuilderView } from "../components/AgentBuilderView";
 import { ConversationView } from "../components/ConversationView";
 import { CreateRoomDialog } from "../components/CreateRoomDialog";
+import { FreeButtonSwitch } from "../components/FreeButtonSwitch";
 import { GroupConversationView } from "../components/GroupConversationView";
 import { InputBar } from "../components/InputBar";
 import { InviteUsersScreen } from "../components/InviteUsersScreen";
+import {
+  PanelLayout,
+  PanelLayoutNav,
+  PanelLayoutPanel,
+} from "../components/PanelLayout";
 import { ProfilePanel } from "../components/Profile";
 import {
   type Agent,
@@ -75,7 +73,6 @@ import {
   createConversationsWithMessages,
   createSpace,
   getAgentById,
-  getConversationsBySpaceId,
   getMembersBySpaceId,
   getRandomAgents,
   getRandomSpaces,
@@ -87,50 +84,58 @@ import {
   type Space,
   type User,
 } from "../data";
+import { getDataSourcesBySpaceId } from "../data/dataSources";
+import { getRandomGreetingForName } from "../data/greetings";
+import {
+  buildPodTabOptions,
+  type DynamicFileTab,
+  getDefaultMainTabOrder,
+  getFileTabValue,
+  type PodTabOption,
+  resolvePodContext,
+  shouldShowMemberChrome,
+} from "./podPanelConfig";
 import TemplateSelection, { type Template } from "./TemplateSelection";
+
+// ── Types ─────────────────────────────────────────────────────────────────────
 
 type Collaborator =
   | { type: "agent"; data: Agent }
   | { type: "person"; data: User };
 
-type Participant =
-  | { type: "user"; data: User }
-  | { type: "agent"; data: Agent };
-
 type SpaceNotificationPreference = "never" | "mentions" | "all";
 
-function getRandomParticipants(conversation: Conversation): Participant[] {
-  const allParticipants: Participant[] = [];
+type PodTabsState = {
+  mainTabOrder: string[];
+  dynamicFileTabs: DynamicFileTab[];
+};
 
-  // Add user participants
-  conversation.userParticipants.forEach((userId) => {
-    const user = getUserById(userId);
-    if (user) {
-      allParticipants.push({ type: "user", data: user });
-    }
-  });
+type SelectedCitation = { title: string; icon?: string };
 
-  // Add agent participants
-  conversation.agentParticipants.forEach((agentId) => {
-    const agent = getAgentById(agentId);
-    if (agent) {
-      allParticipants.push({ type: "agent", data: agent });
-    }
-  });
+// ── Helpers ───────────────────────────────────────────────────────────────────
 
-  // Shuffle and select 1-6 participants
-  const shuffled = [...allParticipants].sort(() => Math.random() - 0.5);
-  const count = Math.min(
-    Math.max(1, Math.floor(Math.random() * 6) + 1),
-    shuffled.length
+function getRandomParticipants(conversation: Conversation) {
+  const all = [
+    ...conversation.userParticipants.map((id) => ({
+      type: "user" as const,
+      data: getUserById(id),
+    })),
+    ...conversation.agentParticipants.map((id) => ({
+      type: "agent" as const,
+      data: getAgentById(id),
+    })),
+  ].filter((p) => p.data != null) as (
+    | { type: "user"; data: User }
+    | { type: "agent"; data: Agent }
+  )[];
+  const shuffled = [...all].sort(() => Math.random() - 0.5);
+  return shuffled.slice(
+    0,
+    Math.min(Math.max(1, Math.floor(Math.random() * 6) + 1), shuffled.length)
   );
-  return shuffled.slice(0, count);
 }
 
-/** Playground-only: treat some conversations as "triggered" for the sidebar Inbox filter. */
-function isPlaygroundInboxTriggeredConversation(
-  conversation: Conversation
-): boolean {
+function isPlaygroundInboxTriggered(conversation: Conversation): boolean {
   let h = 0;
   for (let i = 0; i < conversation.id.length; i++) {
     h = (h + conversation.id.charCodeAt(i)) % 997;
@@ -138,34 +143,95 @@ function isPlaygroundInboxTriggeredConversation(
   return h % 4 === 0;
 }
 
-function DustMain() {
+function getSpaceActivity(space: Space) {
+  const c = space.id.charCodeAt(space.id.length - 1);
+  const count = c % 3 === 0 ? (c % 9) + 1 : undefined;
+  return { count, hasActivity: count ? true : c % 2 !== 0 };
+}
+
+// ── Main component ────────────────────────────────────────────────────────────
+
+function Pods() {
+  // ── Bootstrap state ───────────────────────────────────────────────────────
+  const [user, setUser] = useState<User | null>(null);
+  const [greeting, setGreeting] = useState<string>("");
+  useEffect(() => {
+    if (user) {
+      setGreeting(getRandomGreetingForName(user.firstName));
+    }
+  }, [user]);
+  const [spaces, setSpaces] = useState<Space[]>([]);
+  const [conversationsWithMessages, setConversationsWithMessages] = useState<
+    Conversation[]
+  >([]);
+  const [collaborators, setCollaborators] = useState<Collaborator[]>([]);
+
+  useEffect(() => {
+    const u = getRandomUsers(1)[0];
+    setUser(u);
+    const agentCount = Math.floor(Math.random() * 5) + 1;
+    const peopleCount = Math.floor(Math.random() * 5) + 1;
+    setCollaborators([
+      ...getRandomAgents(agentCount).map((d) => ({
+        type: "agent" as const,
+        data: d,
+      })),
+      ...getRandomUsers(peopleCount).map((d) => ({
+        type: "person" as const,
+        data: d,
+      })),
+    ]);
+    setSpaces(getRandomSpaces(Math.floor(Math.random() * 7) + 3));
+    setConversationsWithMessages(createConversationsWithMessages(u.id));
+  }, []);
+
+  // ── Navigation state ──────────────────────────────────────────────────────
+  // P2 selection: what's shown in the "level 1" panel
+  type P2View =
+    | { kind: "welcome" }
+    | { kind: "conversation"; conversationId: string }
+    | { kind: "space"; spaceId: string }
+    | { kind: "profile" }
+    | { kind: "templates" };
+
+  const [p2View, setP2View] = useState<P2View>({ kind: "welcome" });
+
+  // P3: conversation from a space (level 2) OR citation from a level-1 conversation
+  type P3View =
+    | { kind: "conversation"; conversationId: string }
+    | { kind: "citation"; citation: SelectedCitation };
+
+  const [p3View, setP3View] = useState<P3View | null>(null);
+
+  // P4: citation/attachment opened from a level-2 conversation
+  const [p4Citation, setP4Citation] = useState<SelectedCitation | null>(null);
+
+  // ── Space panel tab state (lifted from GroupConversationView) ────────────
+  const [spaceActiveTab, setSpaceActiveTab] = useState("conversations");
+  const [podTabsBySpaceId, setPodTabsBySpaceId] = useState<
+    Map<string, PodTabsState>
+  >(new Map());
+  const [draggingPodFileId, setDraggingPodFileId] = useState<string | null>(
+    null
+  );
+  const [draggingPodFileName, setDraggingPodFileName] = useState<string | null>(
+    null
+  );
+  const [fileToRevealInKnowledge, setFileToRevealInKnowledge] = useState<
+    string | null
+  >(null);
+
+  // ── Sidebar UI state ──────────────────────────────────────────────────────
   const [activeTab, setActiveTab] = useState<"chat" | "spaces" | "admin">(
     "chat"
   );
   const [searchText, setSearchText] = useState("");
-  const [agentSearchText, setAgentSearchText] = useState("");
-  const [peopleSearchText, setPeopleSearchText] = useState("");
-  const [documentSearchText, setDocumentSearchText] = useState("");
-  const [user, setUser] = useState<User | null>(null);
-  const [collaborators, setCollaborators] = useState<Collaborator[]>([]);
-  const [spaces, setSpaces] = useState<Space[]>([]);
-  const [selectedConversationId, setSelectedConversationId] = useState<
-    string | null
-  >("new-conversation");
-  const [selectedSpaceId, setSelectedSpaceId] = useState<string | null>(null);
-  const [previousSpaceId, setPreviousSpaceId] = useState<string | null>(null);
-  const [selectedView, setSelectedView] = useState<"templates" | null>(null);
-  const [selectedTemplateForBuilder, setSelectedTemplateForBuilder] =
-    useState<Template | null>(null);
-  const [conversationsWithMessages, setConversationsWithMessages] = useState<
-    Conversation[]
-  >([]);
-  const [isCreateRoomDialogOpen, setIsCreateRoomDialogOpen] = useState(false);
-  const [isInviteUsersScreenOpen, setIsInviteUsersScreenOpen] = useState(false);
-  const [lastCreatedSpaceId, setLastCreatedSpaceId] = useState<string | null>(
-    null
-  );
-  const [inviteSpaceId, setInviteSpaceId] = useState<string | null>(null);
+  const [inboxHideTriggered, setInboxHideTriggered] = useState(false);
+  const [isAgentsDropdownOpen, setIsAgentsDropdownOpen] = useState(false);
+  const [spaceNotificationPreferences, setSpaceNotificationPreferences] =
+    useState<Map<string, SpaceNotificationPreference>>(new Map());
+
+  // ── Space management state ────────────────────────────────────────────────
   const [spaceMembers, setSpaceMembers] = useState<Map<string, string[]>>(
     new Map()
   );
@@ -175,259 +241,382 @@ function DustMain() {
   const [spacePublicSettings, setSpacePublicSettings] = useState<
     Map<string, boolean>
   >(new Map());
-  const [spaceNotificationPreferences, setSpaceNotificationPreferences] =
-    useState<Map<string, SpaceNotificationPreference>>(new Map());
+  const [isCreateRoomDialogOpen, setIsCreateRoomDialogOpen] = useState(false);
+  const [isInviteUsersScreenOpen, setIsInviteUsersScreenOpen] = useState(false);
+  const [inviteSpaceId, setInviteSpaceId] = useState<string | null>(null);
+  const [lastCreatedSpaceId, setLastCreatedSpaceId] = useState<string | null>(
+    null
+  );
 
-  // Track sidebar collapsed state for toggle button icon
-  const [isSidebarCollapsed, setIsSidebarCollapsed] = useState(false);
-  const [showProfileView, setShowProfileView] = useState(false);
-  const [isAgentsDropdownOpen, setIsAgentsDropdownOpen] = useState(false);
-  const [inboxHideTriggered, setInboxHideTriggered] = useState(false);
-  const sidebarLayoutRef = useRef<SidebarLayoutRef>(null);
+  // ── Agent builder ─────────────────────────────────────────────────────────
+  const [selectedTemplateForBuilder, setSelectedTemplateForBuilder] =
+    useState<Template | null>(null);
 
-  // Initialize space members with generated members when a space is first selected
+  // Auto-initialize space members
   useEffect(() => {
-    if (selectedSpaceId && !spaceMembers.has(selectedSpaceId)) {
-      const generatedMembers = getMembersBySpaceId(selectedSpaceId);
-      setSpaceMembers((prev) => {
-        const newMap = new Map(prev);
-        newMap.set(selectedSpaceId, generatedMembers);
-        return newMap;
-      });
+    const spaceId = p2View.kind === "space" ? p2View.spaceId : null;
+    if (spaceId && !spaceMembers.has(spaceId)) {
+      setSpaceMembers((prev) =>
+        new Map(prev).set(spaceId, getMembersBySpaceId(spaceId))
+      );
     }
-  }, [selectedSpaceId]); // eslint-disable-line react-hooks/exhaustive-deps
+  }, [p2View, spaceMembers]);
 
-  useEffect(() => {
-    const randomUser = getRandomUsers(1)[0];
-    setUser(randomUser);
-
-    // Generate random total count of collaborators between 2 and 12
-    const totalCount = Math.floor(Math.random() * (12 - 2 + 1)) + 2;
-
-    // Randomly decide how many agents and people
-    const agentCount = Math.floor(Math.random() * (totalCount - 1)) + 1;
-    const peopleCount = totalCount - agentCount;
-
-    const randomAgents = getRandomAgents(agentCount);
-    const randomPeople = getRandomUsers(peopleCount);
-
-    // Create mixed collaborator list
-    const mixedCollaborators: Collaborator[] = [
-      ...randomAgents.map((agent) => ({ type: "agent" as const, data: agent })),
-      ...randomPeople.map((person) => ({
-        type: "person" as const,
-        data: person,
-      })),
-    ];
-
-    setCollaborators(mixedCollaborators);
-
-    // Generate random number of spaces between 3 and 9
-    const spaceCount = Math.floor(Math.random() * (9 - 3 + 1)) + 3;
-    const randomSpaces = getRandomSpaces(spaceCount);
-    setSpaces(randomSpaces);
-
-    // Create conversations with messages
-    const convsWithMessages = createConversationsWithMessages(randomUser.id);
-    setConversationsWithMessages(convsWithMessages);
-  }, []);
-
-  // Auto-select newly created space when it's added to the spaces array
+  // Auto-select newly created space
   useEffect(() => {
     if (lastCreatedSpaceId && spaces.find((s) => s.id === lastCreatedSpaceId)) {
-      setSelectedView(null);
-      setSelectedSpaceId(lastCreatedSpaceId);
-      setSelectedConversationId(null);
+      setP2View({ kind: "space", spaceId: lastCreatedSpaceId });
+      setP3View(null);
       setLastCreatedSpaceId(null);
     }
   }, [spaces, lastCreatedSpaceId]);
 
-  // Combine mock conversations with conversations that have messages
-  const allConversations = useMemo(() => {
-    return [...conversationsWithMessages, ...mockConversations];
-  }, [conversationsWithMessages]);
+  // ── Derived data ──────────────────────────────────────────────────────────
+  const allConversations = useMemo(
+    () => [...conversationsWithMessages, ...mockConversations],
+    [conversationsWithMessages]
+  );
 
   const filteredConversations = useMemo(() => {
-    if (!searchText.trim()) {
-      return allConversations;
-    }
-    const lowerSearch = searchText.toLowerCase();
-    return allConversations.filter((conv) =>
-      conv.title.toLowerCase().includes(lowerSearch)
+    if (!searchText.trim()) return allConversations;
+    const lower = searchText.toLowerCase();
+    return allConversations.filter((c) =>
+      c.title.toLowerCase().includes(lower)
     );
   }, [searchText, allConversations]);
 
   const groupedConversations = useMemo(() => {
-    const today = new Date();
+    const now = new Date();
+    const today = new Date(now);
     today.setHours(0, 0, 0, 0);
-
     const yesterday = new Date(today);
-    yesterday.setDate(yesterday.getDate() - 1);
-
+    yesterday.setDate(today.getDate() - 1);
     const lastWeek = new Date(today);
-    lastWeek.setDate(lastWeek.getDate() - 7);
-
+    lastWeek.setDate(today.getDate() - 7);
     const lastMonth = new Date(today);
-    lastMonth.setDate(lastMonth.getDate() - 30);
-
+    lastMonth.setDate(today.getDate() - 30);
     const groups = {
-      today: [] as typeof filteredConversations,
-      yesterday: [] as typeof filteredConversations,
-      lastWeek: [] as typeof filteredConversations,
-      lastMonth: [] as typeof filteredConversations,
+      today: [] as Conversation[],
+      yesterday: [] as Conversation[],
+      lastWeek: [] as Conversation[],
+      lastMonth: [] as Conversation[],
     };
-
-    filteredConversations.forEach((conv) => {
-      const updatedAt = conv.updatedAt;
-      if (updatedAt >= today) {
-        groups.today.push(conv);
-      } else if (updatedAt >= yesterday) {
-        groups.yesterday.push(conv);
-      } else if (updatedAt >= lastWeek) {
-        groups.lastWeek.push(conv);
-      } else if (updatedAt >= lastMonth) {
-        groups.lastMonth.push(conv);
-      }
+    filteredConversations.forEach((c) => {
+      if (c.updatedAt >= today) groups.today.push(c);
+      else if (c.updatedAt >= yesterday) groups.yesterday.push(c);
+      else if (c.updatedAt >= lastWeek) groups.lastWeek.push(c);
+      else if (c.updatedAt >= lastMonth) groups.lastMonth.push(c);
     });
-
     return groups;
   }, [filteredConversations]);
 
-  const filteredAgents = useMemo(() => {
-    if (!agentSearchText.trim()) {
-      return mockAgents;
-    }
-    const lowerSearch = agentSearchText.toLowerCase();
-    return mockAgents.filter((agent) =>
-      agent.name.toLowerCase().includes(lowerSearch)
-    );
-  }, [agentSearchText]);
-
-  const filteredPeople = useMemo(() => {
-    if (!peopleSearchText.trim()) {
-      return mockUsers;
-    }
-    const lowerSearch = peopleSearchText.toLowerCase();
-    return mockUsers.filter(
-      (person) =>
-        person.fullName.toLowerCase().includes(lowerSearch) ||
-        person.email.toLowerCase().includes(lowerSearch)
-    );
-  }, [peopleSearchText]);
-
-  const sortedCollaborators = useMemo(() => {
-    return [...collaborators].sort((a, b) => {
-      const nameA = a.type === "agent" ? a.data.name : a.data.fullName;
-      const nameB = b.type === "agent" ? b.data.name : b.data.fullName;
-      return nameA.localeCompare(nameB);
-    });
-  }, [collaborators]);
-
-  const filteredCollaborators = useMemo(() => {
-    if (!searchText.trim()) {
-      return sortedCollaborators;
-    }
-    const lowerSearch = searchText.toLowerCase();
-    return sortedCollaborators.filter((collaborator) => {
-      if (collaborator.type === "agent") {
-        const agent = collaborator.data;
-        return (
-          agent.name.toLowerCase().includes(lowerSearch) ||
-          agent.description.toLowerCase().includes(lowerSearch)
-        );
-      } else {
-        const person = collaborator.data;
-        return (
-          person.fullName.toLowerCase().includes(lowerSearch) ||
-          person.email.toLowerCase().includes(lowerSearch)
-        );
-      }
-    });
-  }, [searchText, sortedCollaborators]);
-
-  // Derive count and hasActivity deterministically from space ID.
-  const getSpaceActivity = (space: Space) => {
-    const charCode = space.id.charCodeAt(space.id.length - 1);
-    const count = charCode % 3 === 0 ? (charCode % 9) + 1 : undefined;
-    const hasActivity = count ? true : charCode % 2 !== 0;
-    return { count, hasActivity };
-  };
+  const inboxConversations = useMemo(() => {
+    const pool = inboxHideTriggered
+      ? filteredConversations.filter((c) => !isPlaygroundInboxTriggered(c))
+      : filteredConversations;
+    if (pool.length === 0) return [];
+    const count = Math.floor(Math.random() * 4) + 2;
+    return [...pool]
+      .sort(() => 0.5 - Math.random())
+      .slice(0, Math.min(count, pool.length))
+      .map((c) => ({
+        conversation: c,
+        status: (Math.random() < 0.25 ? "blocked" : "idle") as
+          | "idle"
+          | "blocked",
+      }));
+  }, [filteredConversations, inboxHideTriggered]);
 
   const sortedSpaces = useMemo(() => {
     return [...spaces].sort((a, b) => {
-      const actA = getSpaceActivity(a);
-      const actB = getSpaceActivity(b);
-
-      // 1. Items with count come first, highest count first
-      const countA = actA.count ?? 0;
-      const countB = actB.count ?? 0;
-      if (countA !== countB) {
-        return countB - countA;
-      }
-
-      // 2. Items with hasActivity (but no count) come next
-      if (actA.hasActivity !== actB.hasActivity) {
-        return actA.hasActivity ? -1 : 1;
-      }
-
-      // 3. Alphabetical by name
+      const { count: cA = 0, hasActivity: hA } = getSpaceActivity(a);
+      const { count: cB = 0, hasActivity: hB } = getSpaceActivity(b);
+      if (cA !== cB) return cB - cA;
+      if (hA !== hB) return hA ? -1 : 1;
       return a.name.localeCompare(b.name);
     });
   }, [spaces]);
 
   const filteredSpaces = useMemo(() => {
-    if (!searchText.trim()) {
-      return sortedSpaces;
-    }
-    const lowerSearch = searchText.toLowerCase();
+    if (!searchText.trim()) return sortedSpaces;
+    const lower = searchText.toLowerCase();
     return sortedSpaces.filter(
-      (space) =>
-        space.name.toLowerCase().includes(lowerSearch) ||
-        space.description.toLowerCase().includes(lowerSearch)
+      (s) =>
+        s.name.toLowerCase().includes(lower) ||
+        s.description.toLowerCase().includes(lower)
     );
   }, [searchText, sortedSpaces]);
 
-  // Find selected conversation from all conversations
-  const selectedConversation = useMemo(() => {
-    if (!selectedConversationId) return null;
+  const selectedConversationId =
+    p2View.kind === "conversation" ? p2View.conversationId : null;
+
+  const selectedConversation = useMemo(
+    () =>
+      selectedConversationId
+        ? (allConversations.find((c) => c.id === selectedConversationId) ??
+          null)
+        : null,
+    [selectedConversationId, allConversations]
+  );
+  const p3Conversation = useMemo(
+    () =>
+      p3View?.kind === "conversation"
+        ? (allConversations.find((c) => c.id === p3View.conversationId) ?? null)
+        : null,
+    [p3View, allConversations]
+  );
+
+  // ── Pod context & tab state ───────────────────────────────────────────────
+  const podContext = useMemo(
+    () => resolvePodContext(p2View, spaces, allConversations),
+    [p2View, spaces, allConversations]
+  );
+
+  const activePodTab = spaceActiveTab;
+  const setActivePodTab = setSpaceActiveTab;
+
+  const currentPodTabsState = useMemo((): PodTabsState | null => {
+    if (!podContext) {
+      return null;
+    }
+
     return (
-      allConversations.find((c) => c.id === selectedConversationId) || null
+      podTabsBySpaceId.get(podContext.spaceId) ?? {
+        mainTabOrder: getDefaultMainTabOrder(podContext.variant),
+        dynamicFileTabs: [],
+      }
     );
-  }, [selectedConversationId, allConversations]);
+  }, [podContext, podTabsBySpaceId]);
 
-  // Find selected space
-  const selectedSpace = useMemo(() => {
-    if (!selectedSpaceId) return null;
-    return spaces.find((s) => s.id === selectedSpaceId) || null;
-  }, [selectedSpaceId, spaces]);
+  const basePodTabOptions = useMemo((): PodTabOption[] => {
+    if (!podContext || !currentPodTabsState) {
+      return [];
+    }
 
-  // Get conversations for selected space
-  const spaceConversations = useMemo(() => {
-    if (!selectedSpaceId) return [];
-    return getConversationsBySpaceId(selectedSpaceId);
-  }, [selectedSpaceId]);
+    return buildPodTabOptions(
+      podContext.variant,
+      currentPodTabsState.mainTabOrder,
+      currentPodTabsState.dynamicFileTabs
+    );
+  }, [podContext, currentPodTabsState]);
 
-  // Select 2-5 random conversations for inbox with status assignment
-  const inboxConversations = useMemo(() => {
-    const pool = inboxHideTriggered
-      ? filteredConversations.filter(
-          (c) => !isPlaygroundInboxTriggeredConversation(c)
-        )
-      : filteredConversations;
-    if (pool.length === 0) return [];
+  const dynamicFileTabIds = useMemo(
+    () =>
+      currentPodTabsState?.dynamicFileTabs.map((tab) => tab.dataSourceId) ?? [],
+    [currentPodTabsState]
+  );
 
-    // Randomly select 2-5 conversations
-    const count = Math.floor(Math.random() * 4) + 2; // 2-5
-    const shuffled = [...pool].sort(() => 0.5 - Math.random());
-    const selected = shuffled.slice(0, Math.min(count, pool.length));
+  useEffect(() => {
+    if (!podContext) {
+      return;
+    }
 
-    // Assign statuses: ~25% probability of "blocked", rest "idle"
-    return selected.map((conversation) => {
-      const status: "idle" | "unread" | "blocked" | "error" =
-        Math.random() < 0.25 ? "blocked" : "idle";
-      return { conversation, status };
+    setPodTabsBySpaceId((prev) => {
+      if (prev.has(podContext.spaceId)) {
+        return prev;
+      }
+
+      return new Map(prev).set(podContext.spaceId, {
+        mainTabOrder: getDefaultMainTabOrder(podContext.variant),
+        dynamicFileTabs: [],
+      });
     });
-  }, [filteredConversations, inboxHideTriggered]);
+  }, [podContext]);
+
+  const handlePodTabReorder = useCallback(
+    (nextOptions: PodTabOption[]) => {
+      if (!podContext) {
+        return;
+      }
+
+      const mainTabOrder = nextOptions
+        .filter((option) => option.pinned !== "end")
+        .map((option) => option.value);
+
+      setPodTabsBySpaceId((prev) => {
+        const existing = prev.get(podContext.spaceId) ?? {
+          mainTabOrder: getDefaultMainTabOrder(podContext.variant),
+          dynamicFileTabs: [],
+        };
+
+        return new Map(prev).set(podContext.spaceId, {
+          ...existing,
+          mainTabOrder,
+        });
+      });
+    },
+    [podContext]
+  );
+
+  const handlePodFileDrop = useCallback(
+    (fileId: string) => {
+      if (!podContext) {
+        return;
+      }
+
+      const file = getDataSourcesBySpaceId(podContext.spaceId).find(
+        (dataSource) => dataSource.id === fileId
+      );
+      if (!file || file.kind === "folder") {
+        return;
+      }
+
+      const fileTabValue = getFileTabValue(fileId);
+
+      setPodTabsBySpaceId((prev) => {
+        const existing = prev.get(podContext.spaceId) ?? {
+          mainTabOrder: getDefaultMainTabOrder(podContext.variant),
+          dynamicFileTabs: [],
+        };
+        const alreadyOpen = existing.dynamicFileTabs.some(
+          (tab) => tab.dataSourceId === fileId
+        );
+        const dynamicFileTabs = alreadyOpen
+          ? existing.dynamicFileTabs
+          : [
+              ...existing.dynamicFileTabs,
+              {
+                value: fileTabValue,
+                dataSourceId: fileId,
+                label: file.fileName,
+              },
+            ];
+        const mainTabOrder = alreadyOpen
+          ? existing.mainTabOrder
+          : [...existing.mainTabOrder, fileTabValue];
+
+        return new Map(prev).set(podContext.spaceId, {
+          mainTabOrder,
+          dynamicFileTabs,
+        });
+      });
+      setActivePodTab(fileTabValue);
+      setDraggingPodFileId(null);
+      setDraggingPodFileName(null);
+    },
+    [podContext, setActivePodTab]
+  );
+
+  const handlePodFileDragChange = useCallback(
+    (fileId: string | null, fileName?: string | null) => {
+      setDraggingPodFileId(fileId);
+      setDraggingPodFileName(fileName ?? null);
+    },
+    []
+  );
+
+  const handlePodRemoveTab = useCallback(
+    (tabValue: string) => {
+      if (!podContext || !tabValue.startsWith("file-")) {
+        return;
+      }
+
+      setPodTabsBySpaceId((prev) => {
+        const existing = prev.get(podContext.spaceId);
+        if (!existing) {
+          return prev;
+        }
+
+        return new Map(prev).set(podContext.spaceId, {
+          mainTabOrder: existing.mainTabOrder.filter(
+            (value) => value !== tabValue
+          ),
+          dynamicFileTabs: existing.dynamicFileTabs.filter(
+            (tab) => tab.value !== tabValue
+          ),
+        });
+      });
+
+      if (activePodTab === tabValue) {
+        setActivePodTab("conversations");
+      }
+    },
+    [activePodTab, podContext, setActivePodTab]
+  );
+
+  const handleShowFileInFiles = useCallback(
+    (tabValue: string) => {
+      if (!tabValue.startsWith("file-")) {
+        return;
+      }
+
+      setActivePodTab("knowledge");
+      setFileToRevealInKnowledge(tabValue.slice("file-".length));
+    },
+    [setActivePodTab]
+  );
+
+  const podTabOptions = useMemo((): PodTabOption[] => {
+    return basePodTabOptions.map((option) => {
+      if (!option.value.startsWith("file-")) {
+        return option;
+      }
+
+      return {
+        ...option,
+        contextMenuItems: [
+          {
+            label: "Start a conversation with document",
+            icon: ChatBubbleBottomCenterTextIcon,
+          },
+          {
+            label: "Show in files",
+            icon: EyeIcon,
+            onClick: () => handleShowFileInFiles(option.value),
+          },
+          {
+            label: "Remove from topbar",
+            icon: XMarkIcon,
+            variant: "warning",
+            onClick: () => handlePodRemoveTab(option.value),
+          },
+        ],
+      };
+    });
+  }, [basePodTabOptions, handlePodRemoveTab, handleShowFileInFiles]);
+
+  // ── Handlers ──────────────────────────────────────────────────────────────
+  const handleRoomNameNext = (name: string, isPublic: boolean) => {
+    const newSpace = createSpace(name, undefined, isPublic);
+    setSpaces((prev) => [...prev, newSpace]);
+    setSpacePublicSettings((prev) => new Map(prev).set(newSpace.id, isPublic));
+    setLastCreatedSpaceId(newSpace.id);
+    setIsCreateRoomDialogOpen(false);
+  };
+
+  const handleInviteMembers = (spaceId: string) => {
+    setInviteSpaceId(spaceId);
+    setIsInviteUsersScreenOpen(true);
+  };
+
+  const handleInviteUsersComplete = (
+    selectedUserIds: string[],
+    editorUserIds: string[]
+  ) => {
+    if (inviteSpaceId) {
+      setSpaceMembers((prev) =>
+        new Map(prev).set(inviteSpaceId, selectedUserIds)
+      );
+      setSpaceEditors((prev) =>
+        new Map(prev).set(inviteSpaceId, editorUserIds)
+      );
+    }
+    setIsInviteUsersScreenOpen(false);
+    setInviteSpaceId(null);
+  };
+
+  const handleUpdateSpaceName = (spaceId: string, newName: string) => {
+    setSpaces((prev) =>
+      prev.map((s) => (s.id === spaceId ? { ...s, name: newName } : s))
+    );
+  };
+
+  const handleUpdateSpacePublic = (spaceId: string, isPublic: boolean) => {
+    setSpacePublicSettings((prev) => new Map(prev).set(spaceId, isPublic));
+    setSpaces((prev) =>
+      prev.map((s) => (s.id === spaceId ? { ...s, isPublic } : s))
+    );
+  };
 
   const getConversationMoreMenu = (conversation: Conversation) => {
     const participants = getRandomParticipants(conversation);
@@ -453,32 +642,28 @@ function DustMain() {
             <DropdownMenuPortal>
               <DropdownMenuSubContent>
                 {participants.length > 0 ? (
-                  participants.map((participant) => (
+                  participants.map((p) => (
                     <DropdownMenuItem
                       key={
-                        participant.type === "user"
-                          ? `user-${participant.data.id}`
-                          : `agent-${participant.data.id}`
+                        p.type === "user"
+                          ? `user-${p.data.id}`
+                          : `agent-${p.data.id}`
                       }
-                      label={
-                        participant.type === "user"
-                          ? participant.data.fullName
-                          : participant.data.name
-                      }
+                      label={p.type === "user" ? p.data.fullName : p.data.name}
                       icon={
-                        participant.type === "user" ? (
+                        p.type === "user" ? (
                           <Avatar
                             size="xxs"
-                            name={participant.data.fullName}
-                            visual={participant.data.portrait}
-                            isRounded={true}
+                            name={p.data.fullName}
+                            visual={p.data.portrait}
+                            isRounded
                           />
                         ) : (
                           <Avatar
                             size="xxs"
-                            name={participant.data.name}
-                            emoji={participant.data.emoji}
-                            backgroundColor={participant.data.backgroundColor}
+                            name={p.data.name}
+                            emoji={p.data.emoji}
+                            backgroundColor={p.data.backgroundColor}
                             isRounded={false}
                           />
                         )
@@ -513,45 +698,299 @@ function DustMain() {
 
   if (!user) {
     return (
-      <div className="s-flex s-min-h-screen s-items-center s-justify-center s-bg-background dark:s-bg-background-night">
-        <div className="s-text-center">
-          <p className="s-text-foreground dark:s-text-foreground-night">
-            Loading...
-          </p>
-        </div>
+      <div className="s-flex s-h-screen s-items-center s-justify-center s-bg-background dark:s-bg-background-night">
+        <p className="s-text-foreground dark:s-text-foreground-night">
+          Loading…
+        </p>
       </div>
     );
   }
 
-  // Sidebar content
-  const sidebarContent = (
-    <div className="s-flex s-h-full s-flex-col s-border-r s-border-border s-bg-muted-background dark:s-border-border-night dark:s-bg-muted-background-night">
-      <Tabs
-        value={activeTab}
-        onValueChange={(value) =>
-          setActiveTab(value as "chat" | "spaces" | "admin")
-        }
-        className="s-flex s-min-h-0 s-flex-1 s-flex-col"
-      >
-        <TabsList className="s-mt-3 s-px-2">
-          <TabsTrigger
-            value="chat"
-            label="Chat"
-            icon={ChatBubbleLeftRightIcon}
+  // ── P2 content ────────────────────────────────────────────────────────────
+  const p2Label = (() => {
+    if (podContext) return podContext.space.name;
+    if (p2View.kind === "conversation")
+      return selectedConversation?.title ?? "Conversation";
+    if (p2View.kind === "profile") return "Profile";
+    if (p2View.kind === "templates") return "Templates";
+    return "Home";
+  })();
+
+  const p2Content = (() => {
+    if (p2View.kind === "profile" && user) return <ProfilePanel user={user} />;
+    if (p2View.kind === "templates")
+      return (
+        <div className="s-h-full s-overflow-auto">
+          <TemplateSelection
+            onTemplateClick={(t) => setSelectedTemplateForBuilder(t)}
           />
-          <TabsTrigger value="spaces" label="Spaces" icon={PlanetIcon} />
-          <TabsTrigger value="admin" icon={Cog6ToothIcon} />
-        </TabsList>
-        <TabsContent
-          value="chat"
-          className="s-flex s-min-h-0 s-flex-1 s-flex-col"
-        >
+        </div>
+      );
+    if (p2View.kind === "conversation" && selectedConversation)
+      return (
+        <ConversationView
+          conversation={selectedConversation}
+          locutor={user}
+          users={mockUsers}
+          agents={mockAgents}
+          conversationsWithMessages={conversationsWithMessages}
+          onCitationOpen={(citation) => {
+            setP3View({ kind: "citation", citation });
+            setP4Citation(null);
+          }}
+        />
+      );
+    if (podContext)
+      return (
+        <GroupConversationView
+          space={podContext.space}
+          conversations={podContext.conversations}
+          users={mockUsers}
+          agents={mockAgents}
+          spaceMemberIds={
+            spaceMembers.get(podContext.spaceId) ??
+            getMembersBySpaceId(podContext.spaceId)
+          }
+          editorUserIds={spaceEditors.get(podContext.spaceId) ?? []}
+          onConversationClick={(conversation) => {
+            setP3View({
+              kind: "conversation",
+              conversationId: conversation.id,
+            });
+            setP4Citation(null);
+          }}
+          onInviteMembers={() => handleInviteMembers(podContext.spaceId)}
+          onUpdateSpaceName={handleUpdateSpaceName}
+          onUpdateSpacePublic={handleUpdateSpacePublic}
+          spacePublicSettings={spacePublicSettings}
+          activeTab={activePodTab}
+          onTabChange={setActivePodTab}
+          dynamicFileTabIds={dynamicFileTabIds}
+          onAddFileToTopbar={handlePodFileDrop}
+          onFileDragChange={handlePodFileDragChange}
+          fileToRevealInKnowledge={fileToRevealInKnowledge}
+          onFileToRevealInKnowledgeHandled={() =>
+            setFileToRevealInKnowledge(null)
+          }
+          podVariant={podContext.variant}
+          currentUserId={user.id}
+          selectedConversationId={
+            p3View?.kind === "conversation" ? p3View.conversationId : null
+          }
+        />
+      );
+    // welcome
+    return (
+      <div className="s-flex s-h-full s-w-full s-items-center s-justify-center s-bg-background dark:s-bg-background-night">
+        <div className="s-flex s-w-full s-max-w-4xl s-flex-col s-gap-6 s-px-4 s-py-8">
+          <div className="s-heading-2xl s-text-foreground dark:s-text-foreground-night">
+            {greeting}
+          </div>
+          <InputBar placeholder="Ask a question" />
+          <div className="s-heading-lg s-text-foreground dark:s-text-foreground-night">
+            Chat with…
+          </div>
+        </div>
+      </div>
+    );
+  })();
+
+  // ── P3 content ────────────────────────────────────────────────────────────
+  const p3Label =
+    p3View?.kind === "conversation"
+      ? (p3Conversation?.title ?? "Conversation")
+      : p3View?.kind === "citation"
+        ? p3View.citation.title
+        : "Panel 3";
+
+  const p3Content = (() => {
+    if (!p3View) return null;
+    if (p3View.kind === "conversation" && p3Conversation)
+      return (
+        <ConversationView
+          conversation={p3Conversation}
+          locutor={user}
+          users={mockUsers}
+          agents={mockAgents}
+          conversationsWithMessages={conversationsWithMessages}
+          onCitationOpen={(citation) => setP4Citation(citation)}
+        />
+      );
+    if (p3View.kind === "citation")
+      return (
+        <div className="s-flex s-h-full s-flex-col s-gap-3 s-p-4">
+          <p className="s-text-sm s-font-medium s-text-foreground dark:s-text-foreground-night">
+            {p3View.citation.title}
+          </p>
+          <div className="s-flex-1 s-rounded-lg s-border s-border-separator s-bg-muted-background s-p-4 s-text-sm s-text-muted-foreground dark:s-border-separator-night dark:s-bg-muted-background-night dark:s-text-muted-foreground-night">
+            Document preview placeholder
+          </div>
+        </div>
+      );
+    return null;
+  })();
+
+  // ── P4 content ────────────────────────────────────────────────────────────
+  const p4Label = p4Citation?.title ?? "Attachment";
+  const p4Content = p4Citation ? (
+    <div className="s-flex s-h-full s-flex-col s-gap-3 s-p-4">
+      <p className="s-text-sm s-font-medium s-text-foreground dark:s-text-foreground-night">
+        {p4Citation.title}
+      </p>
+      <div className="s-flex-1 s-rounded-lg s-border s-border-separator s-bg-muted-background s-p-4 s-text-sm s-text-muted-foreground dark:s-border-separator-night dark:s-bg-muted-background-night dark:s-text-muted-foreground-night">
+        Document preview placeholder
+      </div>
+    </div>
+  ) : null;
+
+  // ── Panel top bars ────────────────────────────────────────────────────────
+  const conversationActions = (
+    <>
+      <Button
+        size="sm"
+        variant="ghost-secondary"
+        icon={AttachmentIcon}
+        isSelect
+      />
+      <Button size="sm" variant="ghost-secondary" icon={MoreIcon} />
+    </>
+  );
+
+  const podTopBarLeft = podContext ? (
+    <FreeButtonSwitch
+      value={activePodTab}
+      onValueChange={setActivePodTab}
+      options={podTabOptions}
+      onOptionsReorder={handlePodTabReorder}
+      onDropCreateOption={handlePodFileDrop}
+      onRemoveOption={handlePodRemoveTab}
+      isFileDragActive={draggingPodFileId !== null}
+      draggingFileLabel={draggingPodFileName}
+    />
+  ) : null;
+
+  const podTopBarRight = (() => {
+    if (!podContext || !shouldShowMemberChrome(podContext.variant)) return null;
+    const memberIds =
+      spaceMembers.get(podContext.spaceId) ??
+      getMembersBySpaceId(podContext.spaceId);
+    const memberAvatars = memberIds
+      .map((id) => mockUsers.find((u) => u.id === id))
+      .filter((u): u is (typeof mockUsers)[0] => !!u)
+      .slice(0, 5)
+      .map((u) => ({
+        name: u.fullName,
+        visual: u.portrait,
+        isRounded: true as const,
+      }));
+    return (
+      <div className="s-flex s-items-center s-gap-2">
+        {memberAvatars.length > 0 && (
+          <Avatar.Stack
+            avatars={memberAvatars}
+            nbVisibleItems={memberAvatars.length}
+            orientation="horizontal"
+            hasMagnifier={false}
+            size="sm"
+          />
+        )}
+        <Button
+          size="sm"
+          variant="primary"
+          label="Join"
+          tooltip="Join the project"
+          onClick={() => {}}
+        />
+      </div>
+    );
+  })();
+
+  const p2TopBarLeft = (() => {
+    if (p2View.kind === "conversation" && selectedConversation)
+      return (
+        <Breadcrumbs
+          items={[{ label: selectedConversation.title }]}
+          size="sm"
+          hasLighterFont
+        />
+      );
+    if (podContext) return podTopBarLeft;
+    if (p2View.kind === "profile")
+      return (
+        <Breadcrumbs items={[{ label: "Profile" }]} size="sm" hasLighterFont />
+      );
+    if (p2View.kind === "templates")
+      return (
+        <Breadcrumbs
+          items={[{ label: "Templates" }]}
+          size="sm"
+          hasLighterFont
+        />
+      );
+    return null;
+  })();
+
+  const p2TopBarRight = (() => {
+    if (p2View.kind === "conversation") return conversationActions;
+    if (podContext) return podTopBarRight;
+    return null;
+  })();
+
+  const p3TopBarLeft = (() => {
+    if (p3View?.kind === "conversation" && p3Conversation)
+      return (
+        <Breadcrumbs
+          items={[{ label: p3Conversation.title }]}
+          size="sm"
+          hasLighterFont
+        />
+      );
+    if (p3View?.kind === "citation")
+      return (
+        <Breadcrumbs
+          items={[{ label: p3View.citation.title }]}
+          size="sm"
+          hasLighterFont
+        />
+      );
+    return null;
+  })();
+
+  const p3TopBarRight =
+    p3View?.kind === "conversation" ? conversationActions : null;
+
+  const p4TopBarLeft = p4Citation ? (
+    <Breadcrumbs
+      items={[{ label: p4Citation.title }]}
+      size="sm"
+      hasLighterFont
+    />
+  ) : null;
+
+  // ── Sidebar (Nav) top bar ─────────────────────────────────────────────────
+  const navTopBar = (
+    <FreeButtonSwitch<"chat" | "spaces" | "admin">
+      value={activeTab}
+      onValueChange={setActiveTab}
+      options={[
+        { value: "chat", label: "Chat", icon: ChatBubbleLeftRightIcon },
+        { value: "spaces", label: "Spaces", icon: PlanetIcon },
+        { value: "admin", icon: Cog6ToothIcon },
+      ]}
+    />
+  );
+
+  // ── Sidebar (Nav) content ─────────────────────────────────────────────────
+  const navContent = (
+    <div className="s-flex s-min-h-0 s-flex-1 s-flex-col s-bg-muted-background dark:s-bg-muted-background-night">
+      {/* ── Chat tab ── */}
+      {activeTab === "chat" && (
+        <div className="s-flex s-min-h-0 s-flex-1 s-flex-col">
           <ScrollArea className="s-flex-1">
             <ScrollBar orientation="vertical" size="minimal" />
-            {/* Search Bar */}
-            <div className="s-flex s-gap-2 s-p-2 s-px-2">
+            <div className="s-flex s-gap-2 s-p-2">
               <SearchInput
-                name="conversation-search"
+                name="search"
                 value={searchText}
                 onChange={setSearchText}
                 placeholder="Search"
@@ -570,7 +1009,7 @@ function DustMain() {
               >
                 <DropdownMenuTrigger asChild>
                   <Button
-                    variant="ghost"
+                    variant="ghost-secondary"
                     size="sm"
                     icon={MoreIcon}
                     aria-label="More options"
@@ -603,30 +1042,12 @@ function DustMain() {
                           e.preventDefault();
                           e.stopPropagation();
                           setIsAgentsDropdownOpen(false);
-                          setShowProfileView(false);
-                          setSelectedView("templates");
-                          setSelectedConversationId(null);
-                          setSelectedSpaceId(null);
-                        }}
-                      />
-                      <DropdownMenuItem
-                        label="Open YAML"
-                        icon={CodeSlashIcon}
-                        onClick={(e) => {
-                          e.preventDefault();
-                          e.stopPropagation();
+                          setP2View({ kind: "templates" });
+                          setP3View(null);
                         }}
                       />
                     </DropdownMenuSubContent>
                   </DropdownMenuSub>
-                  <DropdownMenuItem
-                    label="Edit agent"
-                    icon={PencilSquareIcon}
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }}
-                  />
                   <DropdownMenuItem
                     label="Manage agents"
                     icon={ContactsUserIcon}
@@ -656,14 +1077,6 @@ function DustMain() {
                   <DropdownMenuSeparator />
                   <DropdownMenuLabel label="Conversations" />
                   <DropdownMenuItem
-                    label="Edit conversations"
-                    icon={ListSelectIcon}
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }}
-                  />
-                  <DropdownMenuItem
                     label="Clear conversation history"
                     icon={TrashIcon}
                     variant="warning"
@@ -675,6 +1088,8 @@ function DustMain() {
                 </DropdownMenuContent>
               </DropdownMenu>
             </div>
+
+            {/* Inbox */}
             {inboxConversations.length > 0 && (
               <NavigationListCollapsibleSection
                 label="Inbox"
@@ -685,13 +1100,11 @@ function DustMain() {
                     <Button
                       size="xmini"
                       icon={CheckDoubleIcon}
-                      variant="ghost"
-                      aria-label="Mark all as read"
+                      variant="ghost-secondary"
                       tooltip="Mark all as read"
                       onClick={(e) => {
                         e.preventDefault();
                         e.stopPropagation();
-                        // Add action logic here
                       }}
                     />
                     <DropdownMenu>
@@ -699,8 +1112,7 @@ function DustMain() {
                         <Button
                           size="xmini"
                           icon={MoreIcon}
-                          variant="ghost"
-                          aria-label="Inbox options"
+                          variant="ghost-secondary"
                           onClick={(e) => {
                             e.preventDefault();
                             e.stopPropagation();
@@ -708,7 +1120,6 @@ function DustMain() {
                         />
                       </DropdownMenuTrigger>
                       <DropdownMenuContent>
-                        <DropdownMenuLabel label="Conversations" />
                         <DropdownMenuItem
                           label={
                             inboxHideTriggered
@@ -731,22 +1142,27 @@ function DustMain() {
                   <NavigationListItem
                     key={conversation.id}
                     label={conversation.title}
-                    selected={conversation.id === selectedConversationId}
+                    selected={
+                      p2View.kind === "conversation" &&
+                      p2View.conversationId === conversation.id
+                    }
                     status={status}
                     moreMenu={getConversationMoreMenu(conversation)}
                     onClick={() => {
-                      setShowProfileView(false);
-                      setPreviousSpaceId(null);
-                      setSelectedView(null);
-                      setSelectedConversationId(conversation.id);
-                      setSelectedSpaceId(null);
+                      setP2View({
+                        kind: "conversation",
+                        conversationId: conversation.id,
+                      });
+                      setP3View(null);
+                      setP4Citation(null);
                     }}
                   />
                 ))}
               </NavigationListCollapsibleSection>
             )}
-            {/* Collapsible Sections */}
+
             <NavigationList className="s-px-2">
+              {/* Pods */}
               {(filteredSpaces.length > 0 || !searchText.trim()) && (
                 <NavigationListCollapsibleSection
                   label="Pods"
@@ -758,8 +1174,7 @@ function DustMain() {
                       <Button
                         size="xmini"
                         icon={PlusIcon}
-                        variant="ghost"
-                        aria-label="Agents options"
+                        variant="ghost-secondary"
                         onClick={(e) => {
                           e.preventDefault();
                           e.stopPropagation();
@@ -771,8 +1186,7 @@ function DustMain() {
                           <Button
                             size="xmini"
                             icon={MoreIcon}
-                            variant="ghost"
-                            aria-label="Spaces options"
+                            variant="ghost-secondary"
                             onClick={(e) => {
                               e.preventDefault();
                               e.stopPropagation();
@@ -780,14 +1194,6 @@ function DustMain() {
                           />
                         </DropdownMenuTrigger>
                         <DropdownMenuContent>
-                          <DropdownMenuItem
-                            label="Browse"
-                            icon={PencilSquareIcon}
-                            onClick={(e) => {
-                              e.preventDefault();
-                              e.stopPropagation();
-                            }}
-                          />
                           <DropdownMenuItem
                             icon={PlusIcon}
                             label="Create"
@@ -803,20 +1209,20 @@ function DustMain() {
                   }
                 >
                   {filteredSpaces.map((space) => {
-                    // Deterministically assign open or restricted status based on space ID
                     const isRestricted =
                       space.id.charCodeAt(space.id.length - 1) % 2 === 0;
                     const { count, hasActivity } = getSpaceActivity(space);
-                    const spaceMemberIds = getMembersBySpaceId(space.id);
-                    const resolvedSpaceMembers = spaceMemberIds
-                      .map((userId) => getUserById(userId))
-                      .filter((user): user is User => user != null);
+                    const members = getMembersBySpaceId(space.id)
+                      .map((id) => getUserById(id))
+                      .filter((u): u is User => u != null);
                     return (
                       <NavigationListItem
                         key={space.id}
                         label={space.name}
                         icon={isRestricted ? SpaceOpenIcon : SpaceClosedIcon}
-                        selected={space.id === selectedSpaceId}
+                        selected={
+                          p2View.kind === "space" && p2View.spaceId === space.id
+                        }
                         count={count}
                         hasActivity={hasActivity}
                         moreMenu={
@@ -846,18 +1252,14 @@ function DustMain() {
                                         space.id
                                       ) ?? "all"
                                     }
-                                    onValueChange={(value) => {
-                                      setSpaceNotificationPreferences(
-                                        (prev) => {
-                                          const next = new Map(prev);
-                                          next.set(
-                                            space.id,
-                                            value as SpaceNotificationPreference
-                                          );
-                                          return next;
-                                        }
-                                      );
-                                    }}
+                                    onValueChange={(v) =>
+                                      setSpaceNotificationPreferences((prev) =>
+                                        new Map(prev).set(
+                                          space.id,
+                                          v as SpaceNotificationPreference
+                                        )
+                                      )
+                                    }
                                   >
                                     <DropdownMenuRadioItem
                                       value="never"
@@ -899,16 +1301,16 @@ function DustMain() {
                                       handleInviteMembers(space.id);
                                     }}
                                   />
-                                  {resolvedSpaceMembers.map((member) => (
+                                  {members.map((m) => (
                                     <DropdownMenuItem
-                                      key={member.id}
-                                      label={member.fullName}
+                                      key={m.id}
+                                      label={m.fullName}
                                       icon={
                                         <Avatar
-                                          name={member.fullName}
-                                          visual={member.portrait}
+                                          name={m.fullName}
+                                          visual={m.portrait}
                                           size="xxs"
-                                          isRounded={true}
+                                          isRounded
                                         />
                                       }
                                       onClick={(e) => {
@@ -942,10 +1344,9 @@ function DustMain() {
                           </DropdownMenu>
                         }
                         onClick={() => {
-                          setShowProfileView(false);
-                          setSelectedView(null);
-                          setSelectedSpaceId(space.id);
-                          setSelectedConversationId(null);
+                          setP2View({ kind: "space", spaceId: space.id });
+                          setP3View(null);
+                          setP4Citation(null);
                         }}
                       />
                     );
@@ -953,6 +1354,7 @@ function DustMain() {
                 </NavigationListCollapsibleSection>
               )}
 
+              {/* Conversations */}
               {(filteredConversations.length > 0 || !searchText.trim()) && (
                 <NavigationListCollapsibleSection
                   label="Conversations"
@@ -962,8 +1364,7 @@ function DustMain() {
                       <Button
                         size="xmini"
                         icon={ChatBubbleLeftRightIcon}
-                        variant="ghost"
-                        aria-label="New Conversation"
+                        variant="ghost-secondary"
                         tooltip="New Conversation"
                         onClick={(e) => {
                           e.preventDefault();
@@ -975,8 +1376,7 @@ function DustMain() {
                           <Button
                             size="xmini"
                             icon={MoreIcon}
-                            variant="ghost"
-                            aria-label="Conversations options"
+                            variant="ghost-secondary"
                             onClick={(e) => {
                               e.preventDefault();
                               e.stopPropagation();
@@ -984,18 +1384,9 @@ function DustMain() {
                           />
                         </DropdownMenuTrigger>
                         <DropdownMenuContent>
-                          <DropdownMenuLabel label="Conversations" />
                           <DropdownMenuItem
                             label="Hide triggered"
                             icon={BoltOffIcon}
-                            onClick={(e) => {
-                              e.preventDefault();
-                              e.stopPropagation();
-                            }}
-                          />
-                          <DropdownMenuItem
-                            label="Edit history"
-                            icon={ListSelectIcon}
                             onClick={(e) => {
                               e.preventDefault();
                               e.stopPropagation();
@@ -1015,38 +1406,44 @@ function DustMain() {
                     </>
                   }
                 >
-                  {groupedConversations.today.length > 0 && (
-                    <>
-                      {groupedConversations.today.map((conversation) => (
-                        <NavigationListItem
-                          key={conversation.id}
-                          label={conversation.title}
-                          selected={conversation.id === selectedConversationId}
-                          moreMenu={getConversationMoreMenu(conversation)}
-                          onClick={() => {
-                            setShowProfileView(false);
-                            setPreviousSpaceId(null);
-                            setSelectedConversationId(conversation.id);
-                            setSelectedSpaceId(null);
-                          }}
-                        />
-                      ))}
-                    </>
-                  )}
+                  {groupedConversations.today.map((c) => (
+                    <NavigationListItem
+                      key={c.id}
+                      label={c.title}
+                      selected={
+                        p2View.kind === "conversation" &&
+                        p2View.conversationId === c.id
+                      }
+                      moreMenu={getConversationMoreMenu(c)}
+                      onClick={() => {
+                        setP2View({
+                          kind: "conversation",
+                          conversationId: c.id,
+                        });
+                        setP3View(null);
+                        setP4Citation(null);
+                      }}
+                    />
+                  ))}
                   {groupedConversations.yesterday.length > 0 && (
                     <>
                       <NavigationListCompactLabel label="Yesterday" isSticky />
-                      {groupedConversations.yesterday.map((conversation) => (
+                      {groupedConversations.yesterday.map((c) => (
                         <NavigationListItem
-                          key={conversation.id}
-                          label={conversation.title}
-                          selected={conversation.id === selectedConversationId}
-                          moreMenu={getConversationMoreMenu(conversation)}
+                          key={c.id}
+                          label={c.title}
+                          selected={
+                            p2View.kind === "conversation" &&
+                            p2View.conversationId === c.id
+                          }
+                          moreMenu={getConversationMoreMenu(c)}
                           onClick={() => {
-                            setShowProfileView(false);
-                            setPreviousSpaceId(null);
-                            setSelectedConversationId(conversation.id);
-                            setSelectedSpaceId(null);
+                            setP2View({
+                              kind: "conversation",
+                              conversationId: c.id,
+                            });
+                            setP3View(null);
+                            setP4Citation(null);
                           }}
                         />
                       ))}
@@ -1055,17 +1452,22 @@ function DustMain() {
                   {groupedConversations.lastWeek.length > 0 && (
                     <>
                       <NavigationListCompactLabel label="Last week" isSticky />
-                      {groupedConversations.lastWeek.map((conversation) => (
+                      {groupedConversations.lastWeek.map((c) => (
                         <NavigationListItem
-                          key={conversation.id}
-                          label={conversation.title}
-                          selected={conversation.id === selectedConversationId}
-                          moreMenu={getConversationMoreMenu(conversation)}
+                          key={c.id}
+                          label={c.title}
+                          selected={
+                            p2View.kind === "conversation" &&
+                            p2View.conversationId === c.id
+                          }
+                          moreMenu={getConversationMoreMenu(c)}
                           onClick={() => {
-                            setShowProfileView(false);
-                            setPreviousSpaceId(null);
-                            setSelectedConversationId(conversation.id);
-                            setSelectedSpaceId(null);
+                            setP2View({
+                              kind: "conversation",
+                              conversationId: c.id,
+                            });
+                            setP3View(null);
+                            setP4Citation(null);
                           }}
                         />
                       ))}
@@ -1074,17 +1476,22 @@ function DustMain() {
                   {groupedConversations.lastMonth.length > 0 && (
                     <>
                       <NavigationListCompactLabel label="Last month" />
-                      {groupedConversations.lastMonth.map((conversation) => (
+                      {groupedConversations.lastMonth.map((c) => (
                         <NavigationListItem
-                          key={conversation.id}
-                          label={conversation.title}
-                          selected={conversation.id === selectedConversationId}
-                          moreMenu={getConversationMoreMenu(conversation)}
+                          key={c.id}
+                          label={c.title}
+                          selected={
+                            p2View.kind === "conversation" &&
+                            p2View.conversationId === c.id
+                          }
+                          moreMenu={getConversationMoreMenu(c)}
                           onClick={() => {
-                            setShowProfileView(false);
-                            setPreviousSpaceId(null);
-                            setSelectedConversationId(conversation.id);
-                            setSelectedSpaceId(null);
+                            setP2View({
+                              kind: "conversation",
+                              conversationId: c.id,
+                            });
+                            setP3View(null);
+                            setP4Citation(null);
                           }}
                         />
                       ))}
@@ -1094,25 +1501,25 @@ function DustMain() {
               )}
             </NavigationList>
           </ScrollArea>
-        </TabsContent>
-        <TabsContent
-          value="spaces"
-          className="s-flex s-min-h-0 s-flex-1 s-flex-col"
-        >
+        </div>
+      )}
+
+      {activeTab === "spaces" && (
+        <div className="s-flex s-min-h-0 s-flex-1 s-flex-col">
           <div className="s-flex s-flex-1 s-items-center s-justify-center s-text-muted-foreground dark:s-text-muted-foreground-night">
-            Spaces - TBD
+            Spaces — TBD
           </div>
-        </TabsContent>
-        <TabsContent
-          value="admin"
-          className="s-flex s-min-h-0 s-flex-1 s-flex-col"
-        >
+        </div>
+      )}
+      {activeTab === "admin" && (
+        <div className="s-flex s-min-h-0 s-flex-1 s-flex-col">
           <div className="s-flex s-flex-1 s-items-center s-justify-center s-text-muted-foreground dark:s-text-muted-foreground-night">
-            Admin - TBD
+            Admin — TBD
           </div>
-        </TabsContent>
-      </Tabs>
-      {/* Bottom Bar */}
+        </div>
+      )}
+
+      {/* Bottom bar */}
       <div className="s-flex s-h-14 s-items-center s-justify-between s-gap-2 s-border-t s-border-border s-pl-1 s-pr-2 dark:s-border-border-night">
         <DropdownMenu>
           <DropdownMenuTrigger asChild>
@@ -1127,13 +1534,13 @@ function DustMain() {
                   name={user.fullName}
                   visual={user.portrait}
                   size="sm"
-                  isRounded={true}
+                  isRounded
                 />
                 <div className="s-flex s-min-w-0 s-grow s-flex-col s-text-sm s-text-foreground dark:s-text-foreground-night">
-                  <span className="s-heading-sm s-min-w-0 s-overflow-hidden s-text-ellipsis s-whitespace-nowrap">
+                  <span className="s-heading-sm s-truncate">
                     {user.fullName}
                   </span>
-                  <span className="-s-mt-0.5 s-min-w-0 s-overflow-hidden s-text-ellipsis s-whitespace-nowrap s-text-xs s-text-muted-foreground dark:s-text-muted-foreground-night">
+                  <span className="-s-mt-0.5 s-truncate s-text-xs s-text-muted-foreground dark:s-text-muted-foreground-night">
                     ACME
                   </span>
                 </div>
@@ -1147,9 +1554,9 @@ function DustMain() {
               onClick={(e) => {
                 e.preventDefault();
                 e.stopPropagation();
-                setShowProfileView(true);
-                setSelectedConversationId(null);
-                setSelectedSpaceId(null);
+                setP2View({ kind: "profile" });
+                setP3View(null);
+                setP4Citation(null);
               }}
             />
             <DropdownMenuItem
@@ -1164,7 +1571,6 @@ function DustMain() {
               <DropdownMenuSubTrigger icon={HeartIcon} label="Help & Support" />
               <DropdownMenuPortal>
                 <DropdownMenuSubContent>
-                  <DropdownMenuLabel label="Learn about Dust" />
                   <DropdownMenuItem
                     label="Quickstart Guide"
                     icon={LightbulbIcon}
@@ -1174,50 +1580,8 @@ function DustMain() {
                     }}
                   />
                   <DropdownMenuItem
-                    label="Guides & Documentation"
-                    icon={BookOpenIcon}
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }}
-                  />
-                  <DropdownMenuItem
                     label="Join the Slack Community"
                     icon={SlackLogo}
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }}
-                  />
-                  <DropdownMenuLabel label="Ask questions" />
-                  <DropdownMenuItem
-                    label="Ask @help"
-                    description="Ask anything about Dust"
-                    icon={ChatBubbleLeftRightIcon}
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }}
-                  />
-                  <DropdownMenuItem
-                    label="How to invite new users?"
-                    icon={ChatBubbleLeftRightIcon}
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }}
-                  />
-                  <DropdownMenuItem
-                    label="How to use agents in Slack workflow?"
-                    icon={ChatBubbleLeftRightIcon}
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }}
-                  />
-                  <DropdownMenuItem
-                    label="How to manage billing?"
-                    icon={ChatBubbleLeftRightIcon}
                     onClick={(e) => {
                       e.preventDefault();
                       e.stopPropagation();
@@ -1237,201 +1601,64 @@ function DustMain() {
             />
           </DropdownMenuContent>
         </DropdownMenu>
-        <Button
-          variant="ghost-secondary"
-          size="icon"
-          icon={isSidebarCollapsed ? SidebarLeftOpenIcon : SidebarLeftCloseIcon}
-          onClick={() => sidebarLayoutRef.current?.toggle()}
-        />
       </div>
     </div>
   );
 
-  // Handle back button from conversation view
-  const handleConversationBack = () => {
-    setShowProfileView(false);
-    if (previousSpaceId) {
-      setSelectedSpaceId(previousSpaceId);
-      setSelectedConversationId(null);
-      // Optionally clear previousSpaceId, or keep it for future navigation
-      // setPreviousSpaceId(null);
-    }
-  };
-
-  // Handle room creation flow
-  const handleRoomNameNext = (name: string, isPublic: boolean) => {
-    // Create the new space directly
-    const newSpace = createSpace(name, undefined, isPublic);
-
-    // Update spaces state
-    setSpaces((prev) => [...prev, newSpace]);
-
-    // Store the public setting
-    setSpacePublicSettings((prev) => {
-      const newMap = new Map(prev);
-      newMap.set(newSpace.id, isPublic);
-      return newMap;
-    });
-
-    // Track the newly created space ID for auto-selection
-    setLastCreatedSpaceId(newSpace.id);
-
-    // Close dialog
-    setIsCreateRoomDialogOpen(false);
-  };
-
-  // Handle invite members for a space
-  const handleInviteMembers = (spaceId: string) => {
-    setInviteSpaceId(spaceId);
-    setIsInviteUsersScreenOpen(true);
-  };
-
-  const handleInviteUsersComplete = (
-    selectedUserIds: string[],
-    editorUserIds: string[]
-  ) => {
-    // Store invited members for the space
-    if (inviteSpaceId) {
-      setSpaceMembers((prev) => {
-        const newMap = new Map(prev);
-        newMap.set(inviteSpaceId, selectedUserIds);
-        return newMap;
-      });
-      setSpaceEditors((prev) => {
-        const newMap = new Map(prev);
-        newMap.set(inviteSpaceId, editorUserIds);
-        return newMap;
-      });
-    }
-    // Close the invite dialog
-    setIsInviteUsersScreenOpen(false);
-    setInviteSpaceId(null);
-  };
-
-  // Handle space name update
-  const handleUpdateSpaceName = (spaceId: string, newName: string) => {
-    // For prototype, just log the update
-    // In a real implementation, this would update the space in the backend
-    setSpaces((prev) =>
-      prev.map((space) =>
-        space.id === spaceId ? { ...space, name: newName } : space
-      )
-    );
-  };
-
-  // Handle space public setting update
-  const handleUpdateSpacePublic = (spaceId: string, isPublic: boolean) => {
-    // Update the public setting state
-    setSpacePublicSettings((prev) => {
-      const newMap = new Map(prev);
-      newMap.set(spaceId, isPublic);
-      return newMap;
-    });
-    // Also update the space object
-    setSpaces((prev) =>
-      prev.map((space) =>
-        space.id === spaceId ? { ...space, isPublic } : space
-      )
-    );
-    // For prototype, just log the update
-  };
-
-  // Main content
-  const mainContent =
-    // Priority 0: Show profile when opened from user menu
-    showProfileView && user ? (
-      <ProfilePanel user={user} />
-    ) : // Priority 1: Show template selection when Browse templates is clicked
-    selectedView === "templates" ? (
-      <div className="s-h-full s-overflow-auto">
-        <TemplateSelection
-          onTemplateClick={(t) => setSelectedTemplateForBuilder(t)}
-        />
-      </div>
-    ) : // Priority 2: Show conversation view if a conversation is selected (not "new-conversation")
-    selectedConversationId &&
-      selectedConversationId !== "new-conversation" &&
-      selectedConversation &&
-      user ? (
-      <ConversationView
-        conversation={selectedConversation}
-        locutor={user}
-        users={mockUsers}
-        agents={mockAgents}
-        conversationsWithMessages={conversationsWithMessages}
-      />
-    ) : // Priority 3: Show space view if a space is selected
-    selectedSpace && selectedSpaceId ? (
-      <GroupConversationView
-        space={selectedSpace}
-        conversations={spaceConversations}
-        users={mockUsers}
-        agents={mockAgents}
-        spaceMemberIds={
-          spaceMembers.has(selectedSpaceId)
-            ? spaceMembers.get(selectedSpaceId)!
-            : getMembersBySpaceId(selectedSpaceId)
-        }
-        editorUserIds={
-          spaceEditors.has(selectedSpaceId)
-            ? spaceEditors.get(selectedSpaceId)!
-            : []
-        }
-        onConversationClick={(conversation) => {
-          setShowProfileView(false);
-          setSelectedView(null);
-          setPreviousSpaceId(selectedSpaceId);
-          setSelectedConversationId(conversation.id);
-        }}
-        onInviteMembers={() => handleInviteMembers(selectedSpaceId)}
-        onUpdateSpaceName={handleUpdateSpaceName}
-        onUpdateSpacePublic={handleUpdateSpacePublic}
-        spacePublicSettings={spacePublicSettings}
-      />
-    ) : (
-      // Priority 4: Show welcome/new conversation view
-      <div className="s-flex s-h-full s-w-full s-items-center s-justify-center s-bg-background dark:s-bg-background-night">
-        <div className="s-flex s-w-full s-max-w-4xl s-flex-col s-gap-6 s-px-4 s-py-8">
-          <div className="s-heading-2xl s-text-foreground dark:s-text-foreground-night">
-            Welcome, Edouard!{" "}
-          </div>
-          <InputBar />
-          <div className="s-flex s-w-full s-flex-col s-gap-2">
-            <div className="s-heading-lg s-text-foreground dark:s-text-foreground-night">
-              Universal search
-            </div>
-            <SearchInput
-              name="document-search"
-              value={documentSearchText}
-              onChange={setDocumentSearchText}
-              placeholder="Find company documents, Agents, People…"
-              className="s-w-full"
-            />
-          </div>
-          <div className="s-heading-lg s-text-foreground dark:s-text-foreground-night">
-            Chat with…
-          </div>
-        </div>
-      </div>
-    );
-
   return (
-    <div className="s-flex s-h-screen s-w-full s-bg-background dark:s-bg-background-night">
-      <SidebarLayout
-        ref={sidebarLayoutRef}
-        sidebar={sidebarContent}
-        content={mainContent}
-        defaultSidebarWidth={280}
-        minSidebarWidth={200}
-        maxSidebarWidth={400}
-        collapsible={true}
-        onSidebarToggle={setIsSidebarCollapsed}
-      />
+    <>
+      <PanelLayout>
+        <PanelLayoutNav topBarLeft={navTopBar}>
+          {(onNavClose) => (
+            <div
+              className="s-flex s-min-h-0 s-flex-1 s-flex-col"
+              onClick={onNavClose}
+            >
+              {navContent}
+            </div>
+          )}
+        </PanelLayoutNav>
+
+        {/* P2 — Level 1: space, direct conversation, profile, welcome */}
+        <PanelLayoutPanel
+          label={p2Label}
+          isOpen={true}
+          onClose={() => {}}
+          topBarLeft={p2TopBarLeft}
+          topBarRight={p2TopBarRight}
+        >
+          {p2Content}
+        </PanelLayoutPanel>
+
+        {/* P3 — Level 2: conversation from space, or citation from P2 conversation */}
+        <PanelLayoutPanel
+          label={p3Label}
+          isOpen={p3View !== null}
+          onClose={() => {
+            setP3View(null);
+            setP4Citation(null);
+          }}
+          topBarLeft={p3TopBarLeft}
+          topBarRight={p3TopBarRight}
+        >
+          {p3Content}
+        </PanelLayoutPanel>
+
+        {/* P4 — Level 3: citation / attachment */}
+        <PanelLayoutPanel
+          label={p4Label}
+          isOpen={p4Citation !== null}
+          onClose={() => setP4Citation(null)}
+          topBarLeft={p4TopBarLeft}
+        >
+          {p4Content}
+        </PanelLayoutPanel>
+      </PanelLayout>
+
+      {/* Dialogs (outside PanelLayout, portaled to body) */}
       <CreateRoomDialog
         isOpen={isCreateRoomDialogOpen}
-        onClose={() => {
-          setIsCreateRoomDialogOpen(false);
-        }}
+        onClose={() => setIsCreateRoomDialogOpen(false)}
         onNext={handleRoomNameNext}
       />
       <Dialog
@@ -1442,7 +1669,7 @@ function DustMain() {
       >
         <DialogContent
           size="full"
-          className="s-flex s-h-full s-max-h-full s-rounded-none s-p-0 s-overflow-hidden"
+          className="s-flex s-h-full s-max-h-full s-overflow-hidden s-rounded-none s-p-0"
         >
           {selectedTemplateForBuilder && (
             <AgentBuilderView
@@ -1475,10 +1702,10 @@ function DustMain() {
             ? spaceEditors.get(inviteSpaceId)
             : []
         }
-        hasMultipleSelect={true}
+        hasMultipleSelect
       />
-    </div>
+    </>
   );
 }
 
-export default DustMain;
+export default Pods;
